# api/openapi.yaml
openapi: "3.0.3"
info:
  title: ACE-Step Suno-Compatible API
  version: "1.5.0"
  description: |
    100% Suno-API-compatible REST interface powered by ACE-Step 1.5.

    All music-generation endpoints are **asynchronous**: the initial POST returns a
    `taskId` that you poll with the corresponding `record-info` GET until `status`
    reaches `SUCCESS` or `FAILED`.

    Output audio files are served under `/audio/<filename>` as static files.

    ### Authentication
    When the server is configured with `API_KEY`, every request **must** include:
    ```
    Authorization: Bearer <api_key>
    ```
    If `API_KEY` is unset, the header is optional and all requests are accepted.

  contact:
    name: ACE-Step Project
    url: https://github.com/ace-step/ACE-Step-1.5
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:7080
    description: Local development server

# ─────────────────────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: >
        Pass your API key as a Bearer token.
        Leave header absent when the server has no API_KEY configured.

  # ── Shared enums ────────────────────────────────────────────────────────────
  schemas:
...
    GenerateMusicRequest:
      type: object
      required: [customMode, instrumental, model, callBackUrl]
      properties:
        customMode:
          type: boolean
          description: |
            - `true`: user supplies `style` + optionally `prompt` (lyrics).
            - `false`: AI generates both style and lyrics from `prompt`.
        instrumental:
          type: boolean
          description: Generate an instrumental track (no vocals).
        model:
          $ref: "#/components/schemas/ModelVersion"
        callBackUrl:
          type: string
          format: uri
          description: Webhook URL – server POSTs the result here when finished.
        prompt:
          type: string
          description: |
            - `customMode=false`: music description / theme (required).
            - `customMode=true, instrumental=false`: explicit lyrics (required).
        style:
          type: string
          description: Music style tags, e.g. `"Pop, Uplifting, Piano"`. Required when `customMode=true`.
        title:
          type: string
          description: Track title. Required when `customMode=true`.
        personaId:
          type: string
          nullable: true
        personaModel:
          $ref: "#/components/schemas/PersonaModel"
        negativeTags:
          type: string
          nullable: true
          description: Style tags to avoid.
        vocalGender:
          $ref: "#/components/schemas/VocalGender"
        styleWeight:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        weirdnessConstraint:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
        audioWeight:
          type: number
          minimum: 0
          maximum: 1
          nullable: true
...
  responses:
    TaskCreated:
      description: Task accepted; poll `record-info` with the returned `taskId`.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TaskCreatedData"
          example:
            code: 200
            msg: success
            data:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
    TaskDetail:
      description: Current task status and output (if complete).
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TaskDetailData"
    BadRequest:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 400
            msg: "customMode=true requires: style, title"
            data: null
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            msg: "Invalid API key."
            data: null
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/generate:
    post:
      summary: Generate music
      operationId: generate_music
      tags: [Music Generation]
      description: |
        Generate one or more music tracks from a text description.

        **Sync mode** (custom): supply `style` + `prompt` (lyrics), set `thinking=false` internally.

        **AI mode**: supply only `prompt`, let the 5Hz LM write lyrics and style.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMusicRequest"
            examples:
              ai_mode:
                summary: AI-driven (no customMode)
                value:
                  customMode: false
                  instrumental: false
                  model: V4_5PLUS
                  prompt: "An uplifting pop song about finding hope in difficult times"
                  callBackUrl: "https://example.com/webhook"
              custom_mode:
                summary: Custom style + lyrics
                value:
                  customMode: true
                  instrumental: false
                  model: V4_5PLUS
                  style: "Pop, Uplifting, Piano, Female Vocal"
                  title: "Light Ahead"
                  prompt: "[Verse]\nWalking through the storm\nI see a light ahead\n[Chorus]\nHope is all I need"
                  callBackUrl: "https://example.com/webhook"
              instrumental:
                summary: Instrumental track
                value:
                  customMode: true
                  instrumental: true
                  model: V5
                  style: "Cinematic, Orchestral, Epic"
                  title: "Battle Theme"
                  callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/generate/record-info:
    get:
      summary: Poll generation task
      operationId: get_music_generation_details
      tags: [Music Generation]
      parameters:
        - name: taskId
          in: query
          required: true
          schema:
            type: string
          description: Task ID returned by any `/generate` POST endpoint.
          example: a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c
      responses:
        "200":
          $ref: "#/components/responses/TaskDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/generate/extend:
    post:
      summary: Extend an existing track
      operationId: extend_music
      tags: [Music Generation]
      description: |
        Continue a previously generated track from a given time offset.
        Uses ACE-Step `repaint` task internally.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtendMusicRequest"
            example:
              defaultParamFlag: true
              audioId: "d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
              model: V4_5PLUS
              style: "Pop, Uplifting"
              title: "Light Ahead (Extended)"
              prompt: "[Bridge]\nStill moving forward\nNever looking back"
              continueAt: 28.0
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/generate/upload-cover:
    post:
      summary: Cover song with uploaded audio as reference
      operationId: upload_cover
      tags: [Music Generation]
      description: |
        Generates a new track strongly influenced by the style of the
        uploaded reference audio. Uses ACE-Step `cover` task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadCoverRequest"
            example:
              uploadUrl: "https://cdn.example.com/reference.mp3"
              customMode: true
              instrumental: false
              model: V4_5PLUS
              style: "Jazz, Smooth, Trumpet"
              title: "Jazz Cover"
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/generate/upload-extend:
    post:
      summary: Extend an uploaded audio file
      operationId: upload_extend
      tags: [Music Generation]
      description: Upload an MP3/FLAC and extend it from a given time offset.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadExtendRequest"
            example:
              uploadUrl: "https://cdn.example.com/my_song.mp3"
              defaultParamFlag: false
              model: V4_5PLUS
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/generate/mashup:
    post:
      summary: Mashup two audio files
      operationId: generate_mashup
      tags: [Music Generation]
      description: Blend exactly two uploaded audio files into a single track.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMashupRequest"
            example:
              uploadUrlList:
                - "https://cdn.example.com/track_a.mp3"
                - "https://cdn.example.com/track_b.mp3"
              customMode: false
              model: V4_5PLUS
              prompt: "A blend of electronic and classical"
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/generate/add-vocals:
    post:
      summary: Add AI vocals to an instrumental
      operationId: add_vocals
      tags: [Music Generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddVocalsRequest"
            example:
              prompt: "[Verse]\nSunrise over the mountains\nA new day begins"
              title: "Mountain Dawn"
              negativeTags: "Harsh, Metal"
              style: "Folk, Acoustic, Gentle"
              uploadUrl: "https://cdn.example.com/instrumental.mp3"
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/generate/add-instrumental:
    post:
      summary: Add instrumental backing to a vocal stem
      operationId: add_instrumental
      tags: [Music Generation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddInstrumentalRequest"
            example:
              uploadUrl: "https://cdn.example.com/vocal.mp3"
              title: "Backed Vocal"
              negativeTags: "Noise, Heavy"
              tags: "R&B, Smooth, Strings"
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/generate/sounds:
    post:
      summary: Generate ambient sounds / SFX
      operationId: generate_sounds
      tags: [Music Generation]
      description: Generate loopable ambient sounds or short SFX clips.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateSoundsRequest"
            example:
              prompt: "Crackling fireplace with distant rain"
              soundLoop: true
              soundKey: Any
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/generate/get-timestamped-lyrics:
    post:
      summary: Get word-level timestamped lyrics
      operationId: get_timestamped_lyrics
      tags: [Music Generation]
      description: |
        Returns word-aligned timestamps for a completed track.
        Timestamps are best-effort (currently 0.0 for all words unless
        an external aligner is configured).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetTimestampedLyricsRequest"
            example:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
              audioId: "d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
      responses:
        "200":
          description: Timestamped word alignment.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/TimestampedLyricsData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/generate/generate-persona:
    post:
      summary: Create a vocal persona from a track
      operationId: generate_persona
      tags: [Music Generation]
      description: |
        Creates a synthetic persona identifier derived from the specified
        audio segment. The returned `personaId` can be passed back in
        subsequent `/generate` requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePersonaRequest"
            example:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
              audioId: "d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
              name: "Clara"
              description: "Soft, breathy female pop voice"
              vocalStart: 4.0
              vocalEnd: 20.0
      responses:
        "200":
          description: Persona created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PersonaData"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/generate/credit:
    get:
      summary: Remaining credits
      operationId: get_remaining_credits
      tags: [Music Generation]
      description: Returns the simulated credit balance (configured via `SIMULATED_CREDITS`).
      responses:
        "200":
          description: Credit balance.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: integer
                        example: 9999
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Lyrics ──────────────────────────────────────────────────────────────────
...
  /api/v1/vocal-removal/generate:
    post:
      summary: Separate vocals / split stems
      operationId: separate_vocals
      tags: [Vocal Removal]
      description: |
        Runs ACE-Step `extract` task on a previously generated track.
        The `audioId` must reference a track from a completed `GENERATE` task.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeparateVocalsRequest"
            example:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
              audioId: "d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
              type: separate_vocal
              callBackUrl: "https://example.com/webhook"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/wav/generate:
    post:
      summary: Convert audio to WAV
      operationId: convert_to_wav
      tags: [WAV]
      description: |
        Converts a previously generated FLAC/MP3 to 16-bit PCM WAV using
        soundfile (with pydub as fallback).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertToWavRequest"
            example:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
              audioId: "d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/mp4/generate:
    post:
      summary: Generate music video (MP4)
      operationId: create_music_video
      tags: [Video]
      description: |
        Renders a waveform-visualisation MP4 from a previously generated audio track.
        Uses ffmpeg with `showwaves` filter + optional title overlay.
        Falls back to audio-only MP4 if the waveform filter fails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMusicVideoRequest"
            example:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
              audioId: "d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9"
              title: "Light Ahead"
              author: "ACE-Step"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/suno/cover/generate:
    post:
      summary: Generate album cover art
      operationId: generate_cover
      tags: [Cover Art]
      description: |
        Generates a cover image for a completed generation task.
        If `prompt` is omitted, the style tags from the original task are used.

        Image generation priority:
        1. Stable Diffusion (if `diffusers` is installed and GPU available)
        2. PIL gradient placeholder
        3. 1×1 black transparent PNG stub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCoverRequest"
            example:
              taskId: "a3f1b24c8e0d4f7a9b2c1d0e5f6a7b8c"
              callBackUrl: "https://example.com/webhook"
              prompt: "A neon city skyline at dusk, atmospheric, cinematic"
              style: "Synthwave"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/style/generate:
    post:
      summary: Boost style description
      operationId: boost_style
      tags: [Style]
      description: |
        Expands a short style tag string into a richer music production brief
        using the ACE-Step 5Hz LM (`infer_type="dit"`, CoT caption mode).
        Returns the original and expanded versions in the task response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BoostStyleRequest"
            example:
              content: "Dark Synthwave, Cinematic, Tense"
      responses:
        "200":
          $ref: "#/components/responses/TaskCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...